Add "flex", "flexGrow", and "flexShrink" rules. (#5)
Defines some preset "flex", "flexGrow" and "flexShrink" rules and the ability to customize.
Replaces "flexNone" which was outputting "display: none", to output the short hand "flex: none;". "flexAuto" outputs the shorthand "flex: auto;"

Adds in shorthands version of flex-grow: 0 | 1; and flex-shring: 0 | 1; These are the most widely used values for this property, however there are times where we may use a number greater than 1 in order to maintain ratios, so adding the ability to customize "flexGrow" and "flexShrink".

Moves "flex" and "inlineFlex" to be under the "display" rules, which I think makes more sense. Renames them from "flex" to "df" and "inlineFlex" to "dif". Will need to update consuming repos when pull in the latest version of truss.