Don't capture the literal value to avoid inferring never.
In expressions that set the same key (i.e. `color`) to two separate
value, TypeScript would interpret the disjointedness as meaning the
type was impossible to create, and switch it to never.

This wasn't a big deal unles you were trying to object spread the
expression into another properties, as the `...never` would become
a compile error.