Upgrade to Emotion 11 and React 16.14. (#29)
Just housekeeping.

Somewhat odd, I had to add @ts-ignores to the Fela examples that
use Fela's css prop. After digging in to it, I'm suspicious that
the Fela css was ever typed correctly.

In theory I contributed "correct" typings to Fela after noticing
this the first time I played with Fela:

https://github.com/robinweser/fela/pull/792/files

But I think my PR was wrong; that `const fe` typing doesn't have
anything that mentions a `css` key...

But deferring this to a follow up issue.